refactor: move SiteFooter
